Basic yk integration.
This builds CPython with basic yk support. As such yk doesn't do anything: at
the time of this commit, the yk block mapper panics, so we can't even enable a
control point (but this commit does add a commented-out control point as an
example).

Mostly what this commit does is the minimal build integration of yk and CPython
and minimal integration of yk Locations into CPython. Even this is not done
efficiently: every bytecode is given a Location even though very few bytecodes
will ever do anything to a Location. This is clearly wasteful of memory. We
could probably do the same trick as the codeobject's co_opcache which has an
intermediate map of locations (as an array of u8s i.e. a function can have at
most 256 such caches) to save memory. Of course, that's an optimisation rather
than anything fundamental, so we can leave that to a bit later.